requirements.txt
# Core LangChain Dependencies
langchain==0.1.20
langchain-groq==0.1.3
langchain-community==0.0.38
langchain-core==0.1.52

# Groq LLM
groq==0.5.0

# Web Framework
streamlit==1.32.0

# Data Processing
pandas==2.2.1
numpy==1.26.4

# HTTP Requests
requests==2.31.0
httpx==0.27.0

# Environment Management
python-dotenv==1.0.1

# Date and Time
python-dateutil==2.9.0

# Utilities
pydantic==2.7.0
pydantic-settings==2.2.1

# Optional: For better performance
orjson==3.10.0

# Optional: For testing
pytest==8.1.1
pytest-cov==5.0.0

# Optional: Progress bars
tqdm==4.66.2